pyproject.toml
[tool.poetry]
name = "c-core-py"
version = "0.1.0"
description = ""
authors = ["jthetzel <jthetzel@gmail.com>"]
readme = "README.md"
packages = [{include = "c_core_py", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
libgeohash = "^0.1.1"
shapely = "^1.8.5.post1"
pystac = "^1.6.1"
google-cloud-firestore = "^2.7.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
